requirements.txt
faiss-cpu
numpy<2.0
pandas
scikit-learn
pronto
networkx
boto3
matplotlib
seaborn
umap-learn